requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
numpy==1.26.4
packaging==24.1
pygame==2.5.2
PyJWT==1.7.1
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
dj-database-url==2.2.0
psycopg2-binary==2.9.9
setuptools==70.2.0
typing_extensions==4.12.2
tzdata==2024.1
gunicorn==22.0.0